git-svn-id: https://svnsrv.desy.de/desy/EuXFEL/WP76/karabo/karaboFramework/trunk@15664 a23ab9f9-1ea2-4934-960d-990259f9ea8c

+++ build/netbeans/karabo/.bundle.sh
@@ -156,6 +156,14 @@ fi
 cp -rf $DISTDIR/$CONF/$PLATFORM/lib/karathon.so $PYKARABO/ # <-- karathon.so 
 cp -rf $DISTDIR/$CONF/$PLATFORM/include $PACKAGEDIR/
 
+# Check if symbolic link "libkarathon.so" exists. If not, create it! 
+if [ ! -L $PYKARABO/libkarathon.so ]; then 
+    cd $PYKARABO/                       # go to python 'site-packages/karabo' directory
+    ln -s karathon.so libkarathon.so    # create symbolic link
+    cd -                                # return back to previous working directory
+fi
+
+
 # deviceServer
 cd ../deviceServer
 cp -rf $DISTDIR/$CONF/$PLATFORM/bin $PACKAGEDIR/
